I combined user and sp in one object and they will be separated with roll key and delete sp router and sp controller
