+++ Server/Controller/AdController.js
@@ -31,6 +31,7 @@ const Like_Unlike = async (req, res) => {
 };
 
 const CreateAd = async (req, res) => {
+  console.log("in Create ad")
   const {
     location,
     description,
@@ -48,6 +49,7 @@ const CreateAd = async (req, res) => {
     tags,
   } = req.body;
   try {
+    console.log("in try")
     const urls = [];
     req?.files?.forEach((file) => urls.push(file.path));
     const ad = new Ad({
@@ -67,9 +69,11 @@ const CreateAd = async (req, res) => {
       skills_required,
       images: urls,
     });
-    await ad.save();
+    const result=await ad.save();
+    console.log(result)
     res.send({ message: "Ad created successfully!" });
   } catch (error) {
+    console.log("in catch")
     res.status(500).json({ error: error.message });
   }
 };
