+++ Server/Model/UserSchema.js
@@ -1,59 +1,88 @@
-const mongoose=require("mongoose");
-const bcrypt=require("bcryptjs")
-const validator = require('validator');
+const mongoose = require("mongoose");
+const bcrypt = require("bcryptjs");
+const validator = require("validator");
 
-         // ===========================  User  Schema  ===========================
-
-const userSchema=new mongoose.Schema({
-    email:{
-        type:String,
-        required:true,
-        lowercase: true,
-        unique: true,
-        validate: [validator.isEmail, 'Please provide a valid email']
-
+const userSchema = new mongoose.Schema(
+  {
+    email: {
+      type: String,
+      required: true,
+      lowercase: true,
+      unique: true,
+      validate: [validator.isEmail, "Please provide a valid email"],
     },
-    roll:{
-        type:Number
+    roll: {
+      type: Number,
     },
-    username:{
-        type:String,
-        required:true
+    title: {
+      type: String,
     },
-    phone:{
-        type:String,
-        required:true
+    description: {
+      type: String,
     },
-    password:{
-        type:String,
-        required:true,
-        minlength: [8, 'Password should be at least 8 characters long']
+    per_hour: {
+      type: Number,
     },
-    profile_image:{
-        type:String
+    location: {
+      type: String,
     },
-    total_jobs:[{
-        type:Object               
-    }],
-    rating:{
-        type:Number,
-    }
-    
-},{timestamps:true})
+    username: {
+      type: String,
+      required: true,
+    },
+    name: {
+      type: String
+    },
+    phone: {
+      type: String,
+      required: true,
+    },
+    password: {
+      type: String,
+      required: true,
+      minlength: [8, "Password should be at least 8 characters long"],
+    },
+    total_earnings: {
+      type: Number,
+      default: 0,
+    },
+    total_jobs: [
+      {
+        type: Object,
+      },
+    ],
+    total_hours: {
+      type: Number,
+      default: 0,
+    },
+    skills: [
+      {
+        type: String,
+      },
+    ],
+    education: [
+      {
+        type: String,
+      },
+    ],
+    experience: [
+      {
+        type: String,
+      },
+    ],
+    profile_image: {
+      type: String,
+    },
+  },
+  { timestamps: true }
+);
 
+userSchema.pre("save", async function (next) {
+  if (this.isModified("password")) {
+    this.password = await bcrypt.hash(this.password, 10);
+    next();
+  }
+});
 
-userSchema.pre("save", async function(next){
-
-    if (this.isModified("password")){
-
-        this.password=await bcrypt.hash(this.password,10)
-        
-        next()
-    }
-})
-
-const User=mongoose.model("User",userSchema)
-
-
-        // =========================  Export module  =============================
-module.exports=User;
+const User = mongoose.model("User", userSchema);
+module.exports = User;
