+++ Server/server.js
@@ -17,7 +17,6 @@ require("./DB/Connection")
 // ============================================      IMPORT ROUTERS      ===============================
 
 const UserRouter=require("./Router/UserRouter")
-const SPRouter=require("./Router/SPRouter")
 const AdRouter=require("./Router/AdRouter")
 
 
@@ -27,7 +26,6 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 
 app.use(UserRouter)
-app.use(SPRouter)
 app.use(AdRouter)
